backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
sqlalchemy>=2.0.0
python-jose[cryptography]>=3.3.0   
python-multipart>=0.0.9             
httpx>=0.27.0                      
python-dotenv>=1.0.0
google-genai>=1.0.0
pymupdf>=1.24.0                     
stripe>=9.0.0